Update vo_vpdau.c to return - in reconfig().
VO_VDPAU: mpv exits after VT switch out

When the VT switch out is triggered, the decode thread (VD) falls back to sw decoding. However, on the VO thread, which is responsible for handling display preemption and presentation, vo_vdpau.c:reconfig() is called. The reconfig() function returns -1 when the check_preemption() returns 0. vo_reconfig2() returns -1 which enters into an error handling path. This leads to a series of functions calls that ultimately sets the in_terminate flag to TRUE. This leads the vo_thread to exit which ultimately exits the MPV application.

The fix is to return 0 instead of -1 after the check_preemption(). Since the display has been preempted, it is okay to drop the frames and continue.